refactor server execution services and passthrough semantics
