+++ internal/app/app.go
@@ -197,23 +197,24 @@ func New(ctx context.Context, cfg Config) (*App, error) {
 	// Create server
 	allowPassthroughV1Alias := appCfg.Server.AllowPassthroughV1Alias
 	serverCfg := &server.Config{
-		MasterKey:                   appCfg.Server.MasterKey,
-		MetricsEnabled:              appCfg.Metrics.Enabled,
-		MetricsEndpoint:             appCfg.Metrics.Endpoint,
-		BodySizeLimit:               appCfg.Server.BodySizeLimit,
-		AuditLogger:                 auditResult.Logger,
-		UsageLogger:                 usageResult.Logger,
-		PricingResolver:             providerResult.Registry,
-		ModelResolver:               app.aliases.Service,
-		TranslatedRequestPatcher:    translatedRequestPatcher,
-		BatchRequestPreparer:        batchRequestPreparer,
-		ExposedModelLister:          app.aliases.Service,
-		BatchStore:                  batchResult.Store,
-		LogOnlyModelInteractions:    appCfg.Logging.OnlyModelInteractions,
-		DisablePassthroughRoutes:    !appCfg.Server.EnablePassthroughRoutes,
-		EnabledPassthroughProviders: appCfg.Server.EnabledPassthroughProviders,
-		AllowPassthroughV1Alias:     &allowPassthroughV1Alias,
-		SwaggerEnabled:              appCfg.Server.SwaggerEnabled,
+		MasterKey:                    appCfg.Server.MasterKey,
+		MetricsEnabled:               appCfg.Metrics.Enabled,
+		MetricsEndpoint:              appCfg.Metrics.Endpoint,
+		BodySizeLimit:                appCfg.Server.BodySizeLimit,
+		AuditLogger:                  auditResult.Logger,
+		UsageLogger:                  usageResult.Logger,
+		PricingResolver:              providerResult.Registry,
+		ModelResolver:                app.aliases.Service,
+		TranslatedRequestPatcher:     translatedRequestPatcher,
+		BatchRequestPreparer:         batchRequestPreparer,
+		ExposedModelLister:           app.aliases.Service,
+		PassthroughSemanticEnrichers: cfg.Factory.PassthroughSemanticEnrichers(),
+		BatchStore:                   batchResult.Store,
+		LogOnlyModelInteractions:     appCfg.Logging.OnlyModelInteractions,
+		DisablePassthroughRoutes:     !appCfg.Server.EnablePassthroughRoutes,
+		EnabledPassthroughProviders:  appCfg.Server.EnabledPassthroughProviders,
+		AllowPassthroughV1Alias:      &allowPassthroughV1Alias,
+		SwaggerEnabled:               appCfg.Server.SwaggerEnabled,
 	}
 
 	// Initialize admin API and dashboard (behind separate feature flags)
