+++ internal/responsecache/middleware_test.go
@@ -146,6 +146,46 @@ func TestSimpleCacheMiddleware_SkipsStreaming(t *testing.T) {
 	}
 }
 
+func TestSimpleCacheMiddleware_SkipsPartialTranslatedPlan(t *testing.T) {
+	store := cache.NewMapStore()
+	defer store.Close()
+	mw := NewResponseCacheMiddlewareWithStore(store, time.Hour)
+	e := echo.New()
+	e.Use(func(next echo.HandlerFunc) echo.HandlerFunc {
+		return func(c *echo.Context) error {
+			desc := core.DescribeEndpoint(c.Request().Method, c.Request().URL.Path)
+			ctx := core.WithExecutionPlan(c.Request().Context(), &core.ExecutionPlan{
+				Endpoint:     desc,
+				Mode:         core.ExecutionModeTranslated,
+				Capabilities: core.CapabilitiesForEndpoint(desc),
+			})
+			c.SetRequest(c.Request().WithContext(ctx))
+			return next(c)
+		}
+	})
+	e.Use(mw.Middleware())
+	callCount := 0
+	e.POST("/v1/chat/completions", func(c *echo.Context) error {
+		callCount++
+		return c.JSON(http.StatusOK, map[string]string{"n": "1"})
+	})
+
+	body := []byte(`{"model":"gpt-4","messages":[{"role":"user","content":"hi"}]}`)
+	for i := 0; i < 2; i++ {
+		req := httptest.NewRequest(http.MethodPost, "/v1/chat/completions", bytes.NewReader(body))
+		req.Header.Set("Content-Type", "application/json")
+		rec := httptest.NewRecorder()
+		e.ServeHTTP(rec, req)
+		if rec.Header().Get("X-Cache") != "" {
+			t.Fatalf("partial translated plan should bypass cache, got X-Cache=%q", rec.Header().Get("X-Cache"))
+		}
+	}
+
+	if callCount != 2 {
+		t.Fatalf("partial translated plans should bypass cache, handler called %d times", callCount)
+	}
+}
+
 func TestIsStreamingRequest(t *testing.T) {
 	tests := []struct {
 		name string
