+++ internal/server/handlers.go
@@ -2,20 +2,9 @@
 package server
 
 import (
-	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
 	"io"
-	"log/slog"
-	"math"
 	"net/http"
-	"sort"
-	"strconv"
-	"strings"
-	"time"
 
-	"github.com/google/uuid"
 	"github.com/labstack/echo/v5"
 
 	"gomodel/internal/auditlog"
@@ -23,12 +12,6 @@ import (
 	"gomodel/internal/core"
 	"gomodel/internal/usage"
 )
-
-var batchResultsPending404Providers = map[string]struct{}{
-	"anthropic": {},
-}
-
-var defaultEnabledPassthroughProviders = []string{"openai", "anthropic"}
 
 // Handler holds the HTTP handlers
 type Handler struct {
@@ -80,418 +63,47 @@ func (h *Handler) SetBatchStore(store batchstore.Store) {
 	h.batchStore = store
 }
 
-func (h *Handler) setEnabledPassthroughProviders(providerTypes []string) {
-	h.enabledPassthroughProviders = normalizeEnabledPassthroughProviders(providerTypes)
-}
-
 // handleStreamingResponse handles SSE streaming responses for both ChatCompletion and Responses endpoints.
 // It wraps the stream with audit logging and usage tracking, and sets appropriate SSE headers.
 func (h *Handler) handleStreamingResponse(c *echo.Context, model, provider string, streamFn func() (io.ReadCloser, error)) error {
-	// Call streamFn first - only mark as streaming after success
-	// This ensures failed streams are logged normally by handleError/middleware
-	stream, err := streamFn()
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	// Mark as streaming so middleware doesn't log (StreamLogWrapper handles it)
-	auditlog.MarkEntryAsStreaming(c, true)
-	auditlog.EnrichEntryWithStream(c, true)
-
-	// Get entry from context and wrap stream for logging
-	entry := auditlog.GetStreamEntryFromContext(c)
-	streamEntry := auditlog.CreateStreamEntry(entry)
-	if streamEntry != nil {
-		streamEntry.StatusCode = http.StatusOK // Streaming always starts with 200 OK
-	}
-	wrappedStream := auditlog.WrapStreamForLogging(stream, h.logger, streamEntry, c.Request().URL.Path)
-
-	// Wrap with usage tracking if enabled
-	requestID := requestIDFromContextOrHeader(c.Request())
-	endpoint := c.Request().URL.Path
-	wrappedStream = usage.WrapStreamForUsage(wrappedStream, h.usageLogger, model, provider, requestID, endpoint, h.pricingResolver)
-
-	defer func() {
-		_ = wrappedStream.Close() //nolint:errcheck
-	}()
-
-	c.Response().Header().Set("Content-Type", "text/event-stream")
-	c.Response().Header().Set("Cache-Control", "no-cache")
-	c.Response().Header().Set("Connection", "keep-alive")
-
-	// Capture response headers on stream entry AFTER setting them
-	if streamEntry != nil && streamEntry.Data != nil {
-		streamEntry.Data.ResponseHeaders = map[string]string{
-			"Content-Type":  "text/event-stream",
-			"Cache-Control": "no-cache",
-			"Connection":    "keep-alive",
-		}
-	}
-
-	c.Response().WriteHeader(http.StatusOK)
-	if err := flushStream(c.Response(), wrappedStream); err != nil {
-		recordStreamingError(streamEntry, model, provider, c.Request().URL.Path, requestID, err)
-	}
-	return nil
+	return h.translatedInference().handleStreamingResponse(c, model, provider, streamFn)
 }
 
-func flushStream(w io.Writer, stream io.Reader) error {
-	flusher, canFlush := w.(http.Flusher)
-	if canFlush {
-		flusher.Flush()
-	}
-
-	buf := make([]byte, 32*1024)
-	for {
-		n, err := stream.Read(buf)
-		if n > 0 {
-			if _, writeErr := w.Write(buf[:n]); writeErr != nil {
-				return writeErr
-			}
-			if canFlush {
-				flusher.Flush()
-			}
-		}
-		if err != nil {
-			if err == io.EOF {
-				return nil
-			}
-			return err
-		}
+func (h *Handler) translatedInference() *translatedInferenceService {
+	return &translatedInferenceService{
+		provider:                 h.provider,
+		modelResolver:            h.modelResolver,
+		translatedRequestPatcher: h.translatedRequestPatcher,
+		logger:                   h.logger,
+		usageLogger:              h.usageLogger,
+		pricingResolver:          h.pricingResolver,
 	}
 }
 
-func recordStreamingError(streamEntry *auditlog.LogEntry, model, provider, path, requestID string, err error) {
-	if streamEntry != nil {
-		streamEntry.ErrorType = "stream_error"
-		if streamEntry.Data == nil {
-			streamEntry.Data = &auditlog.LogData{}
-		}
-		streamEntry.Data.ErrorMessage = err.Error()
-	}
-
-	slog.Warn("stream terminated abnormally",
-		"error", err,
-		"model", model,
-		"provider", provider,
-		"path", path,
-		"request_id", requestID,
-	)
-}
-
-func requestIDFromContextOrHeader(req *http.Request) string {
-	if req == nil {
-		return ""
-	}
-	requestID := strings.TrimSpace(core.GetRequestID(req.Context()))
-	if requestID != "" {
-		return requestID
-	}
-	return strings.TrimSpace(req.Header.Get("X-Request-ID"))
-}
-
-func requestContextWithRequestID(req *http.Request) (context.Context, string) {
-	if req == nil {
-		requestID := uuid.NewString()
-		return core.WithRequestID(context.Background(), requestID), requestID
-	}
-
-	requestID := requestIDFromContextOrHeader(req)
-	if requestID == "" {
-		requestID = uuid.NewString()
-	}
-
-	req.Header.Set("X-Request-ID", requestID)
-
-	ctx := req.Context()
-	if strings.TrimSpace(core.GetRequestID(ctx)) != requestID {
-		ctx = core.WithRequestID(ctx, requestID)
-		*req = *req.WithContext(ctx)
-	}
-
-	return ctx, requestID
-}
-
-func sanitizePublicBatchMetadata(metadata map[string]string) map[string]string {
-	if len(metadata) == 0 {
-		return nil
-	}
-
-	publicMetadata := make(map[string]string, len(metadata))
-	for key, value := range metadata {
-		switch key {
-		case batchstore.RequestIDMetadataKey, batchstore.UsageLoggedAtMetadataKey:
-			continue
-		default:
-			publicMetadata[key] = value
-		}
-	}
-	if len(publicMetadata) == 0 {
-		return nil
-	}
-	return publicMetadata
-}
-
-func (h *Handler) logUsage(model, providerType string, extractFn func(*core.ModelPricing) *usage.UsageEntry) {
-	if h.usageLogger == nil || !h.usageLogger.Config().Enabled {
-		return
-	}
-	var pricing *core.ModelPricing
-	if h.pricingResolver != nil {
-		pricing = h.pricingResolver.ResolvePricing(model, providerType)
-	}
-	if entry := extractFn(pricing); entry != nil {
-		h.usageLogger.Write(entry)
+func (h *Handler) nativeBatch() *nativeBatchService {
+	return &nativeBatchService{
+		provider:                             h.provider,
+		modelResolver:                        h.modelResolver,
+		batchRequestPreparer:                 h.batchRequestPreparer,
+		batchStore:                           h.batchStore,
+		loadBatch:                            h.loadBatch,
+		cleanupPreparedBatchInputFile:        h.cleanupPreparedBatchInputFile,
+		cleanupStoredBatchRewrittenInputFile: h.cleanupStoredBatchRewrittenInputFile,
+		logBatchUsageFromBatchResults:        h.logBatchUsageFromBatchResults,
 	}
 }
 
-func (h *Handler) shouldEnforceReturningUsageData() bool {
-	return h.usageLogger != nil && h.usageLogger.Config().EnforceReturningUsageData
+func (h *Handler) nativeFiles() *nativeFileService {
+	return &nativeFileService{provider: h.provider}
 }
 
-func cloneChatRequestForStreamUsage(req *core.ChatRequest) *core.ChatRequest {
-	if req == nil {
-		return nil
+func (h *Handler) passthrough() *passthroughService {
+	return &passthroughService{
+		provider:                     h.provider,
+		logger:                       h.logger,
+		normalizePassthroughV1Prefix: h.normalizePassthroughV1Prefix,
+		enabledPassthroughProviders:  h.enabledPassthroughProviders,
 	}
-	cloned := *req
-	if req.StreamOptions != nil {
-		streamOptions := *req.StreamOptions
-		cloned.StreamOptions = &streamOptions
-	}
-	return &cloned
-}
-
-func isEnabledPassthroughProvider(providerType string, enabledPassthroughProviders map[string]struct{}) bool {
-	providerType = strings.TrimSpace(providerType)
-	if providerType == "" {
-		return false
-	}
-	_, ok := enabledPassthroughProviders[providerType]
-	return ok
-}
-
-func normalizeEnabledPassthroughProviders(providerTypes []string) map[string]struct{} {
-	allowed := make(map[string]struct{}, len(providerTypes))
-	for _, providerType := range providerTypes {
-		providerType = strings.TrimSpace(providerType)
-		if providerType == "" {
-			continue
-		}
-		allowed[providerType] = struct{}{}
-	}
-	return allowed
-}
-
-func (h *Handler) enabledPassthroughProviderNames() []string {
-	providers := make([]string, 0, len(h.enabledPassthroughProviders))
-	for providerType := range h.enabledPassthroughProviders {
-		providers = append(providers, providerType)
-	}
-	sort.Strings(providers)
-	return providers
-}
-
-func (h *Handler) unsupportedPassthroughProviderError(providerType string) error {
-	providers := h.enabledPassthroughProviderNames()
-	if len(providers) == 0 {
-		return core.NewInvalidRequestError("provider passthrough is not enabled for any providers", nil)
-	}
-	return core.NewInvalidRequestError(
-		fmt.Sprintf("provider passthrough for %q is not enabled; currently enabled providers: %s", strings.TrimSpace(providerType), strings.Join(providers, ", ")),
-		nil,
-	)
-}
-
-func normalizePassthroughEndpoint(endpoint string, enabled bool) (string, error) {
-	endpoint = strings.TrimSpace(endpoint)
-	switch {
-	case endpoint == "v1":
-		if !enabled {
-			return "", core.NewInvalidRequestError("provider passthrough v1 alias is disabled; use /p/{provider}/... without the v1 prefix", nil)
-		}
-		return "", nil
-	case strings.HasPrefix(endpoint, "v1/"):
-		if !enabled {
-			return "", core.NewInvalidRequestError("provider passthrough v1 alias is disabled; use /p/{provider}/... without the v1 prefix", nil)
-		}
-		return strings.TrimPrefix(endpoint, "v1/"), nil
-	default:
-		return endpoint, nil
-	}
-}
-
-func (h *Handler) passthroughEndpoint(c *echo.Context) (string, string, error) {
-	providerType, endpoint, ok := core.ParseProviderPassthroughPath(c.Request().URL.Path)
-	if !ok {
-		return "", "", core.NewInvalidRequestError("invalid provider passthrough path", nil)
-	}
-	endpoint, err := normalizePassthroughEndpoint(endpoint, h.normalizePassthroughV1Prefix)
-	if err != nil {
-		return "", "", err
-	}
-	if endpoint == "" {
-		return "", "", core.NewInvalidRequestError("provider passthrough endpoint is required", nil)
-	}
-	if rawQuery := strings.TrimSpace(c.Request().URL.RawQuery); rawQuery != "" {
-		endpoint += "?" + rawQuery
-	}
-	return providerType, endpoint, nil
-}
-
-func buildPassthroughHeaders(ctx context.Context, src http.Header) http.Header {
-	connectionHeaders := passthroughConnectionHeaders(src)
-	dst := make(http.Header)
-	for key, values := range src {
-		canonicalKey := http.CanonicalHeaderKey(strings.TrimSpace(key))
-		if skipPassthroughRequestHeader(canonicalKey) || len(values) == 0 {
-			continue
-		}
-		if _, hopByHop := connectionHeaders[canonicalKey]; hopByHop {
-			continue
-		}
-		clonedValues := make([]string, len(values))
-		copy(clonedValues, values)
-		dst[canonicalKey] = clonedValues
-	}
-	requestID := strings.TrimSpace(src.Get("X-Request-ID"))
-	if requestID == "" {
-		requestID = strings.TrimSpace(core.GetRequestID(ctx))
-	}
-	if requestID != "" && strings.TrimSpace(dst.Get("X-Request-ID")) == "" {
-		dst.Set("X-Request-ID", requestID)
-	}
-	if len(dst) == 0 {
-		return nil
-	}
-	return dst
-}
-
-func skipPassthroughHeader(key string) bool {
-	canonicalKey := http.CanonicalHeaderKey(strings.TrimSpace(key))
-	switch canonicalKey {
-	case "Authorization", "X-Api-Key", "Host", "Content-Length", "Connection", "Keep-Alive",
-		"Proxy-Authenticate", "Proxy-Authorization", "Te", "Trailer", "Transfer-Encoding", "Upgrade",
-		"Cookie", "Forwarded", "Set-Cookie":
-		return true
-	default:
-		return strings.HasPrefix(canonicalKey, "X-Forwarded-")
-	}
-}
-
-func skipPassthroughRequestHeader(key string) bool {
-	return skipPassthroughHeader(key)
-}
-
-func passthroughConnectionHeaders(headers http.Header) map[string]struct{} {
-	var tokens map[string]struct{}
-	for key, values := range headers {
-		if http.CanonicalHeaderKey(strings.TrimSpace(key)) != "Connection" {
-			continue
-		}
-		for _, value := range values {
-			for _, token := range strings.Split(value, ",") {
-				canonicalKey := http.CanonicalHeaderKey(strings.TrimSpace(token))
-				if canonicalKey == "" {
-					continue
-				}
-				if tokens == nil {
-					tokens = make(map[string]struct{})
-				}
-				tokens[canonicalKey] = struct{}{}
-			}
-		}
-	}
-	return tokens
-}
-
-func copyPassthroughResponseHeaders(dst, src http.Header) {
-	connectionHeaders := passthroughConnectionHeaders(src)
-	for key, values := range src {
-		canonicalKey := http.CanonicalHeaderKey(strings.TrimSpace(key))
-		if skipPassthroughHeader(canonicalKey) || len(values) == 0 {
-			continue
-		}
-		if _, hopByHop := connectionHeaders[canonicalKey]; hopByHop {
-			continue
-		}
-		dst.Del(canonicalKey)
-		for _, value := range values {
-			dst.Add(canonicalKey, value)
-		}
-	}
-}
-
-func isSSEContentType(headers map[string][]string) bool {
-	for key, values := range headers {
-		if !strings.EqualFold(key, "Content-Type") {
-			continue
-		}
-		for _, value := range values {
-			if strings.Contains(strings.ToLower(value), "text/event-stream") {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func passthroughStreamAuditPath(requestPath, providerType, endpoint string) string {
-	normalized := "/" + strings.TrimLeft(strings.SplitN(endpoint, "?", 2)[0], "/")
-	switch providerType {
-	case "openai":
-		switch normalized {
-		case "/chat/completions":
-			return "/v1/chat/completions"
-		case "/responses":
-			return "/v1/responses"
-		}
-	case "anthropic":
-		switch normalized {
-		case "/messages":
-			return "/v1/messages"
-		}
-	}
-	return requestPath
-}
-
-func (h *Handler) proxyPassthroughResponse(c *echo.Context, providerType, endpoint string, resp *core.PassthroughResponse) error {
-	if resp == nil || resp.Body == nil {
-		return handleError(c, core.NewProviderError(providerType, http.StatusBadGateway, "provider returned empty passthrough response", nil))
-	}
-	defer func() {
-		_ = resp.Body.Close()
-	}()
-
-	copyPassthroughResponseHeaders(c.Response().Header(), http.Header(resp.Headers))
-
-	if isSSEContentType(resp.Headers) {
-		auditlog.MarkEntryAsStreaming(c, true)
-		auditlog.EnrichEntryWithStream(c, true)
-
-		entry := auditlog.GetStreamEntryFromContext(c)
-		streamEntry := auditlog.CreateStreamEntry(entry)
-		if streamEntry != nil {
-			streamEntry.StatusCode = resp.StatusCode
-		}
-
-		wrappedStream := auditlog.WrapStreamForLogging(resp.Body, h.logger, streamEntry, passthroughStreamAuditPath(c.Request().URL.Path, providerType, endpoint))
-		defer func() {
-			_ = wrappedStream.Close()
-		}()
-
-		c.Response().WriteHeader(resp.StatusCode)
-		if err := flushStream(c.Response(), wrappedStream); err != nil {
-			recordStreamingError(streamEntry, "", providerType, c.Request().URL.Path, requestIDFromContextOrHeader(c.Request()), err)
-			return err
-		}
-		return nil
-	}
-
-	c.Response().WriteHeader(resp.StatusCode)
-	if _, err := io.Copy(c.Response(), resp.Body); err != nil {
-		return err
-	}
-	return nil
 }
 
 // ProviderPassthrough handles opaque provider-native requests under /p/{provider}/{endpoint}.
@@ -525,39 +137,7 @@ func (h *Handler) proxyPassthroughResponse(c *echo.Context, providerType, endpoi
 // @Router       /p/{provider}/{endpoint} [head]
 // @Router       /p/{provider}/{endpoint} [options]
 func (h *Handler) ProviderPassthrough(c *echo.Context) error {
-	passthroughProvider, ok := h.provider.(core.RoutablePassthrough)
-	if !ok {
-		return handleError(c, core.NewInvalidRequestError("provider passthrough is not supported by the current provider router", nil))
-	}
-
-	providerType, endpoint, err := h.passthroughEndpoint(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-	if !isEnabledPassthroughProvider(providerType, h.enabledPassthroughProviders) {
-		return handleError(c, h.unsupportedPassthroughProviderError(providerType))
-	}
-
-	ctx := c.Request().Context()
-	requestID := strings.TrimSpace(c.Request().Header.Get("X-Request-ID"))
-	if requestID == "" {
-		requestID = strings.TrimSpace(core.GetRequestID(ctx))
-	}
-	if requestID != "" {
-		ctx = core.WithRequestID(ctx, requestID)
-	}
-	resp, err := passthroughProvider.Passthrough(ctx, providerType, &core.PassthroughRequest{
-		Method:   c.Request().Method,
-		Endpoint: endpoint,
-		Body:     c.Request().Body,
-		Headers:  buildPassthroughHeaders(ctx, c.Request().Header),
-	})
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	auditlog.EnrichEntry(c, "passthrough", providerType)
-	return h.proxyPassthroughResponse(c, providerType, endpoint, resp)
+	return h.passthrough().ProviderPassthrough(c)
 }
 
 // ChatCompletion handles POST /v1/chat/completions
@@ -576,56 +156,7 @@ func (h *Handler) ProviderPassthrough(c *echo.Context) error {
 // @Failure      502      {object}  core.GatewayError
 // @Router       /v1/chat/completions [post]
 func (h *Handler) ChatCompletion(c *echo.Context) error {
-	req, err := canonicalJSONRequestFromSemantics[*core.ChatRequest](c, core.DecodeChatRequest)
-	if err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-	plan, err := ensureTranslatedRequestPlan(c, h.provider, h.modelResolver, &req.Model, &req.Provider)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	ctx := c.Request().Context()
-	if h.translatedRequestPatcher != nil {
-		req, err = h.translatedRequestPatcher.PatchChatRequest(ctx, req)
-		if err != nil {
-			return handleError(c, err)
-		}
-	}
-	providerType := GetProviderType(c)
-	if plan != nil && strings.TrimSpace(plan.ProviderType) != "" {
-		providerType = plan.ProviderType
-	}
-	requestID := requestIDFromContextOrHeader(c.Request())
-
-	if req.Stream {
-		streamReq := req
-		if h.shouldEnforceReturningUsageData() {
-			streamReq = cloneChatRequestForStreamUsage(req)
-			if streamReq.StreamOptions == nil {
-				streamReq.StreamOptions = &core.StreamOptions{}
-			}
-			streamReq.StreamOptions.IncludeUsage = true
-		}
-		usageModel := streamReq.Model
-		if plan != nil && plan.Resolution != nil && strings.TrimSpace(plan.Resolution.ResolvedSelector.Model) != "" {
-			usageModel = plan.Resolution.ResolvedSelector.Model
-		}
-		return h.handleStreamingResponse(c, usageModel, providerType, func() (io.ReadCloser, error) {
-			return h.provider.StreamChatCompletion(ctx, streamReq)
-		})
-	}
-
-	resp, err := h.provider.ChatCompletion(ctx, req)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	h.logUsage(resp.Model, providerType, func(pricing *core.ModelPricing) *usage.UsageEntry {
-		return usage.ExtractFromChatResponse(resp, requestID, providerType, "/v1/chat/completions", pricing)
-	})
-
-	return c.JSON(http.StatusOK, resp)
+	return h.translatedInference().ChatCompletion(c)
 }
 
 // Health handles GET /health
@@ -665,117 +196,6 @@ func (h *Handler) ListModels(c *echo.Context) error {
 	return c.JSON(http.StatusOK, resp)
 }
 
-func (h *Handler) nativeFileRouter() (core.NativeFileRoutableProvider, error) {
-	nativeRouter, ok := h.provider.(core.NativeFileRoutableProvider)
-	if !ok {
-		return nil, core.NewInvalidRequestError("file routing is not supported by the current provider router", nil)
-	}
-	return nativeRouter, nil
-}
-
-func (h *Handler) fileProviderTypes(ctx *echo.Context) ([]string, error) {
-	typed, ok := h.provider.(core.NativeFileProviderTypeLister)
-	if !ok {
-		return nil, core.NewProviderError("", http.StatusInternalServerError, "file provider inventory is unavailable", nil)
-	}
-	return typed.NativeFileProviderTypes(), nil
-}
-
-func (h *Handler) fileByID(
-	c *echo.Context,
-	callFn func(core.NativeFileRoutableProvider, string, string) (any, error),
-	respondFn func(*echo.Context, any) error,
-) error {
-	nativeRouter, err := h.nativeFileRouter()
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	fileReq, err := fileRouteInfoFromSemantics(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	id := strings.TrimSpace(fileReq.FileID)
-	if id == "" {
-		return handleError(c, core.NewInvalidRequestError("file id is required", nil))
-	}
-
-	if providerType := fileReq.Provider; providerType != "" {
-		auditlog.EnrichEntry(c, "file", providerType)
-		result, err := callFn(nativeRouter, providerType, id)
-		if err != nil {
-			return handleError(c, err)
-		}
-		return respondFn(c, result)
-	}
-
-	providers, err := h.fileProviderTypes(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-	auditlog.EnrichEntry(c, "file", "")
-
-	var firstErr error
-	for _, candidate := range providers {
-		result, err := callFn(nativeRouter, candidate, id)
-		if err == nil {
-			return respondFn(c, result)
-		}
-		if isNotFoundGatewayError(err) || isUnsupportedNativeFilesError(err) {
-			continue
-		}
-		if firstErr == nil {
-			firstErr = err
-		}
-	}
-	if firstErr != nil {
-		return handleError(c, firstErr)
-	}
-	return handleError(c, core.NewNotFoundError("file not found: "+id))
-}
-
-func isNotFoundGatewayError(err error) bool {
-	var gatewayErr *core.GatewayError
-	return errors.As(err, &gatewayErr) && gatewayErr.HTTPStatusCode() == http.StatusNotFound
-}
-
-func isUnsupportedNativeFilesError(err error) bool {
-	var gatewayErr *core.GatewayError
-	if !errors.As(err, &gatewayErr) {
-		return false
-	}
-	if gatewayErr.HTTPStatusCode() != http.StatusBadRequest {
-		return false
-	}
-	return strings.Contains(strings.ToLower(gatewayErr.Message), "does not support native file operations")
-}
-
-func sortFilesDesc(items []core.FileObject) {
-	sort.Slice(items, func(i, j int) bool {
-		if items[i].CreatedAt == items[j].CreatedAt {
-			return items[i].ID > items[j].ID
-		}
-		return items[i].CreatedAt > items[j].CreatedAt
-	})
-}
-
-func applyAfterCursor(items []core.FileObject, after string) ([]core.FileObject, error) {
-	after = strings.TrimSpace(after)
-	if after == "" {
-		return items, nil
-	}
-	for i := range items {
-		if items[i].ID == after {
-			if i+1 >= len(items) {
-				return []core.FileObject{}, nil
-			}
-			return items[i+1:], nil
-		}
-	}
-	return nil, core.NewNotFoundError("after cursor file not found: " + after)
-}
-
 // CreateFile handles POST /v1/files.
 //
 // @Summary      Upload a file
@@ -792,70 +212,7 @@ func applyAfterCursor(items []core.FileObject, after string) ([]core.FileObject,
 // @Failure      502       {object}  core.GatewayError
 // @Router       /v1/files [post]
 func (h *Handler) CreateFile(c *echo.Context) error {
-	nativeRouter, err := h.nativeFileRouter()
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	fileReq, err := fileRouteInfoFromSemantics(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	providers, err := h.fileProviderTypes(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	providerType := fileReq.Provider
-	if providerType == "" {
-		if len(providers) == 1 {
-			providerType = providers[0]
-		} else if len(providers) == 0 {
-			return handleError(c, core.NewInvalidRequestError("no providers are available for file uploads", nil))
-		} else {
-			return handleError(c, core.NewInvalidRequestError("provider is required when multiple providers are configured; pass ?provider=<type>", nil))
-		}
-	}
-	auditlog.EnrichEntry(c, "file", providerType)
-
-	purpose := strings.TrimSpace(fileReq.Purpose)
-	if purpose == "" {
-		return handleError(c, core.NewInvalidRequestError("purpose is required", nil))
-	}
-
-	fileHeader, err := c.FormFile("file")
-	if err != nil {
-		return handleError(c, core.NewInvalidRequestError("file is required", err))
-	}
-	file, err := fileHeader.Open()
-	if err != nil {
-		return handleError(c, core.NewInvalidRequestError("failed to open uploaded file", err))
-	}
-	defer func() {
-		_ = file.Close()
-	}()
-
-	content, err := io.ReadAll(file)
-	if err != nil {
-		return handleError(c, core.NewInvalidRequestError("failed to read uploaded file", err))
-	}
-
-	requestID := strings.TrimSpace(c.Request().Header.Get("X-Request-ID"))
-	ctx := core.WithRequestID(c.Request().Context(), requestID)
-	filename := strings.TrimSpace(fileReq.Filename)
-	if filename == "" {
-		filename = fileHeader.Filename
-	}
-	resp, err := nativeRouter.CreateFile(ctx, providerType, &core.FileCreateRequest{
-		Purpose:  purpose,
-		Filename: filename,
-		Content:  content,
-	})
-	if err != nil {
-		return handleError(c, err)
-	}
-	return c.JSON(http.StatusOK, resp)
+	return h.nativeFiles().CreateFile(c)
 }
 
 // ListFiles handles GET /v1/files.
@@ -875,90 +232,7 @@ func (h *Handler) CreateFile(c *echo.Context) error {
 // @Failure      502       {object}  core.GatewayError
 // @Router       /v1/files [get]
 func (h *Handler) ListFiles(c *echo.Context) error {
-	nativeRouter, err := h.nativeFileRouter()
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	fileReq, err := fileRouteInfoFromSemantics(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-	limit := 20
-	if fileReq.HasLimit {
-		limit = fileReq.Limit
-	}
-	if limit <= 0 {
-		limit = 20
-	}
-	if limit > 100 {
-		limit = 100
-	}
-
-	purpose := fileReq.Purpose
-	after := fileReq.After
-	providerType := fileReq.Provider
-
-	if providerType != "" {
-		auditlog.EnrichEntry(c, "file", providerType)
-		resp, err := nativeRouter.ListFiles(c.Request().Context(), providerType, purpose, limit, after)
-		if err != nil {
-			return handleError(c, err)
-		}
-		if resp == nil {
-			resp = &core.FileListResponse{Object: "list"}
-		}
-		if resp.Object == "" {
-			resp.Object = "list"
-		}
-		return c.JSON(http.StatusOK, resp)
-	}
-
-	providers, err := h.fileProviderTypes(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-	auditlog.EnrichEntry(c, "file", "")
-
-	aggregated := make([]core.FileObject, 0)
-	anySuccess := false
-	var firstErr error
-	for _, candidate := range providers {
-		resp, err := nativeRouter.ListFiles(c.Request().Context(), candidate, purpose, limit+1, "")
-		if err != nil {
-			if isUnsupportedNativeFilesError(err) || isNotFoundGatewayError(err) {
-				continue
-			}
-			if firstErr == nil {
-				firstErr = err
-			}
-			continue
-		}
-		anySuccess = true
-		if resp == nil {
-			continue
-		}
-		aggregated = append(aggregated, resp.Data...)
-	}
-	if !anySuccess && firstErr != nil {
-		return handleError(c, firstErr)
-	}
-
-	sortFilesDesc(aggregated)
-	aggregated, err = applyAfterCursor(aggregated, after)
-	if err != nil {
-		return handleError(c, err)
-	}
-	hasMore := len(aggregated) > limit
-	if hasMore {
-		aggregated = aggregated[:limit]
-	}
-
-	return c.JSON(http.StatusOK, core.FileListResponse{
-		Object:  "list",
-		Data:    aggregated,
-		HasMore: hasMore,
-	})
+	return h.nativeFiles().ListFiles(c)
 }
 
 // GetFile handles GET /v1/files/{id}.
@@ -976,14 +250,7 @@ func (h *Handler) ListFiles(c *echo.Context) error {
 // @Failure      502       {object}  core.GatewayError
 // @Router       /v1/files/{id} [get]
 func (h *Handler) GetFile(c *echo.Context) error {
-	return h.fileByID(c,
-		func(r core.NativeFileRoutableProvider, provider, id string) (any, error) {
-			return r.GetFile(c.Request().Context(), provider, id)
-		},
-		func(c *echo.Context, result any) error {
-			return c.JSON(http.StatusOK, result)
-		},
-	)
+	return h.nativeFiles().GetFile(c)
 }
 
 // DeleteFile handles DELETE /v1/files/{id}.
@@ -1001,14 +268,7 @@ func (h *Handler) GetFile(c *echo.Context) error {
 // @Failure      502       {object}  core.GatewayError
 // @Router       /v1/files/{id} [delete]
 func (h *Handler) DeleteFile(c *echo.Context) error {
-	return h.fileByID(c,
-		func(r core.NativeFileRoutableProvider, provider, id string) (any, error) {
-			return r.DeleteFile(c.Request().Context(), provider, id)
-		},
-		func(c *echo.Context, result any) error {
-			return c.JSON(http.StatusOK, result)
-		},
-	)
+	return h.nativeFiles().DeleteFile(c)
 }
 
 // GetFileContent handles GET /v1/files/{id}/content.
@@ -1026,22 +286,7 @@ func (h *Handler) DeleteFile(c *echo.Context) error {
 // @Failure      502       {object}  core.GatewayError
 // @Router       /v1/files/{id}/content [get]
 func (h *Handler) GetFileContent(c *echo.Context) error {
-	return h.fileByID(c,
-		func(r core.NativeFileRoutableProvider, provider, id string) (any, error) {
-			return r.GetFileContent(c.Request().Context(), provider, id)
-		},
-		func(c *echo.Context, result any) error {
-			resp, ok := result.(*core.FileContentResponse)
-			if !ok || resp == nil {
-				return handleError(c, core.NewProviderError("", http.StatusBadGateway, "provider returned empty file content response", nil))
-			}
-			contentType := strings.TrimSpace(resp.ContentType)
-			if contentType == "" {
-				contentType = "application/octet-stream"
-			}
-			return c.Blob(http.StatusOK, contentType, resp.Data)
-		},
-	)
+	return h.nativeFiles().GetFileContent(c)
 }
 
 // Responses handles POST /v1/responses
@@ -1059,51 +304,7 @@ func (h *Handler) GetFileContent(c *echo.Context) error {
 // @Failure      502      {object}  core.GatewayError
 // @Router       /v1/responses [post]
 func (h *Handler) Responses(c *echo.Context) error {
-	req, err := canonicalJSONRequestFromSemantics[*core.ResponsesRequest](c, core.DecodeResponsesRequest)
-	if err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-	plan, err := ensureTranslatedRequestPlan(c, h.provider, h.modelResolver, &req.Model, &req.Provider)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	ctx := c.Request().Context()
-	if h.translatedRequestPatcher != nil {
-		req, err = h.translatedRequestPatcher.PatchResponsesRequest(ctx, req)
-		if err != nil {
-			return handleError(c, err)
-		}
-	}
-	providerType := GetProviderType(c)
-	if plan != nil && strings.TrimSpace(plan.ProviderType) != "" {
-		providerType = plan.ProviderType
-	}
-	requestID := requestIDFromContextOrHeader(c.Request())
-
-	if req.Stream {
-		if h.shouldEnforceReturningUsageData() {
-			ctx = core.WithEnforceReturningUsageData(ctx, true)
-		}
-		usageModel := req.Model
-		if plan != nil && plan.Resolution != nil && strings.TrimSpace(plan.Resolution.ResolvedSelector.Model) != "" {
-			usageModel = plan.Resolution.ResolvedSelector.Model
-		}
-		return h.handleStreamingResponse(c, usageModel, providerType, func() (io.ReadCloser, error) {
-			return h.provider.StreamResponses(ctx, req)
-		})
-	}
-
-	resp, err := h.provider.Responses(ctx, req)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	h.logUsage(resp.Model, providerType, func(pricing *core.ModelPricing) *usage.UsageEntry {
-		return usage.ExtractFromResponsesResponse(resp, requestID, providerType, "/v1/responses", pricing)
-	})
-
-	return c.JSON(http.StatusOK, resp)
+	return h.translatedInference().Responses(c)
 }
 
 // Embeddings handles POST /v1/embeddings
@@ -1121,32 +322,7 @@ func (h *Handler) Responses(c *echo.Context) error {
 // @Failure      502      {object}  core.GatewayError
 // @Router       /v1/embeddings [post]
 func (h *Handler) Embeddings(c *echo.Context) error {
-	req, err := canonicalJSONRequestFromSemantics[*core.EmbeddingRequest](c, core.DecodeEmbeddingRequest)
-	if err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-	plan, err := ensureTranslatedRequestPlan(c, h.provider, h.modelResolver, &req.Model, &req.Provider)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	ctx := c.Request().Context()
-	providerType := GetProviderType(c)
-	if plan != nil && strings.TrimSpace(plan.ProviderType) != "" {
-		providerType = plan.ProviderType
-	}
-	requestID := requestIDFromContextOrHeader(c.Request())
-
-	resp, err := h.provider.Embeddings(ctx, req)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	h.logUsage(resp.Model, providerType, func(pricing *core.ModelPricing) *usage.UsageEntry {
-		return usage.ExtractFromEmbeddingResponse(resp, requestID, providerType, "/v1/embeddings", pricing)
-	})
-
-	return c.JSON(http.StatusOK, resp)
+	return h.translatedInference().Embeddings(c)
 }
 
 // Batches handles POST /v1/batches.
@@ -1166,213 +342,7 @@ func (h *Handler) Embeddings(c *echo.Context) error {
 // @Failure      502      {object}  core.GatewayError
 // @Router       /v1/batches [post]
 func (h *Handler) Batches(c *echo.Context) error {
-	req, err := canonicalJSONRequestFromSemantics[*core.BatchRequest](c, core.DecodeBatchRequest)
-	if err != nil {
-		return handleError(c, core.NewInvalidRequestError("invalid request body: "+err.Error(), err))
-	}
-
-	ctx, requestID := requestContextWithRequestID(c.Request())
-	batchPreparation := &core.BatchPreparationMetadata{}
-	ctx = core.WithBatchPreparationMetadata(ctx, batchPreparation)
-
-	nativeRouter, ok := h.provider.(core.NativeBatchRoutableProvider)
-	if !ok {
-		return handleError(c, core.NewInvalidRequestError("batch routing is not supported by the current provider router", nil))
-	}
-
-	providerType, err := determineBatchProviderType(h.provider, h.modelResolver, req)
-	if err != nil {
-		return handleError(c, err)
-	}
-	auditlog.EnrichEntry(c, "batch", providerType)
-
-	forward := req
-	var preparedHints map[string]string
-	if h.batchRequestPreparer != nil {
-		prepared, err := h.batchRequestPreparer.PrepareBatchRequest(ctx, providerType, req)
-		if err != nil {
-			return handleError(c, err)
-		}
-		if prepared != nil {
-			if prepared.Request != nil {
-				forward = prepared.Request
-			}
-			batchPreparation.RecordInputFileRewrite(prepared.OriginalInputFileID, prepared.RewrittenInputFileID)
-			preparedHints = prepared.RequestEndpointHints
-		}
-	}
-
-	var (
-		upstream *core.BatchResponse
-		hints    map[string]string
-	)
-	if hintedRouter, ok := h.provider.(core.NativeBatchHintRoutableProvider); ok {
-		upstream, hints, err = hintedRouter.CreateBatchWithHints(ctx, providerType, forward)
-	} else {
-		upstream, err = nativeRouter.CreateBatch(ctx, providerType, forward)
-	}
-	if err != nil {
-		h.cleanupPreparedBatchInputFile(ctx, providerType, batchPreparation.RewrittenInputFileID)
-		return handleError(c, err)
-	}
-	if upstream == nil {
-		return handleError(c, core.NewProviderError(providerType, http.StatusBadGateway, "provider returned empty batch response", nil))
-	}
-
-	providerBatchID := upstream.ProviderBatchID
-	if providerBatchID == "" {
-		providerBatchID = upstream.ID
-	}
-	if providerBatchID == "" {
-		return handleError(c, core.NewProviderError(providerType, http.StatusBadGateway, "provider response missing batch id", nil))
-	}
-
-	resp := *upstream
-	resp.Provider = providerType
-	resp.ProviderBatchID = providerBatchID
-	resp.ID = "batch_" + uuid.NewString()
-	resp.Object = "batch"
-	resp.InputFileID = firstNonEmpty(req.InputFileID, batchPreparation.OriginalInputFileID, resp.InputFileID)
-	if resp.Endpoint == "" {
-		resp.Endpoint = core.NormalizeOperationPath(req.Endpoint)
-	}
-	if resp.CompletionWindow == "" {
-		resp.CompletionWindow = req.CompletionWindow
-	}
-	if resp.CompletionWindow == "" {
-		resp.CompletionWindow = "24h"
-	}
-	if resp.Metadata == nil {
-		resp.Metadata = map[string]string{}
-	}
-	resp.Metadata["provider"] = providerType
-	resp.Metadata["provider_batch_id"] = providerBatchID
-	resp.Metadata = sanitizePublicBatchMetadata(resp.Metadata)
-
-	if h.batchStore != nil {
-		mergedHints := mergeBatchRequestEndpointHints(preparedHints, hints)
-		stored := &batchstore.StoredBatch{
-			Batch:                     &resp,
-			RequestEndpointByCustomID: mergedHints,
-			OriginalInputFileID:       batchPreparation.OriginalInputFileID,
-			RewrittenInputFileID:      batchPreparation.RewrittenInputFileID,
-			RequestID:                 requestID,
-		}
-		if err := h.batchStore.Create(ctx, stored); err != nil {
-			return handleError(c, core.NewProviderError("batch_store", http.StatusInternalServerError, "failed to persist batch", err))
-		}
-		if hintedRouter, ok := h.provider.(core.NativeBatchHintRoutableProvider); ok && len(mergedHints) > 0 {
-			hintedRouter.ClearBatchResultHints(providerType, providerBatchID)
-		}
-	}
-
-	return c.JSON(http.StatusOK, resp)
-}
-
-func determineBatchProviderType(provider core.RoutableProvider, resolver RequestModelResolver, req *core.BatchRequest) (string, error) {
-	if provider == nil {
-		return "", core.NewInvalidRequestError("provider is not configured", nil)
-	}
-
-	if strings.TrimSpace(req.InputFileID) != "" {
-		if req.Metadata == nil {
-			return "", core.NewInvalidRequestError("metadata.provider is required for input_file_id batches", nil)
-		}
-		providerType := strings.TrimSpace(req.Metadata["provider"])
-		if providerType == "" {
-			return "", core.NewInvalidRequestError("metadata.provider is required for input_file_id batches", nil)
-		}
-		return providerType, nil
-	}
-
-	if len(req.Requests) == 0 {
-		return "", core.NewInvalidRequestError("requests is required and must not be empty", nil)
-	}
-
-	var providerType string
-	resolver = effectiveRequestModelResolver(provider, resolver)
-	for i, item := range req.Requests {
-		selector, err := core.BatchItemModelSelector(req.Endpoint, item)
-		if err != nil {
-			return "", core.NewInvalidRequestError(fmt.Sprintf("batch item %d: %s", i, err.Error()), err)
-		}
-		if resolver != nil {
-			selector, _, err = resolver.ResolveModel(selector.Model, selector.Provider)
-			if err != nil {
-				return "", core.NewInvalidRequestError(fmt.Sprintf("batch item %d: %s", i, err.Error()), err)
-			}
-		}
-		model := selector.QualifiedModel()
-		if model == "" {
-			return "", core.NewInvalidRequestError(fmt.Sprintf("batch item %d: model is required", i), nil)
-		}
-		if !provider.Supports(model) {
-			return "", core.NewInvalidRequestError("unsupported model: "+model, nil)
-		}
-		itemProvider := provider.GetProviderType(model)
-		if providerType == "" {
-			providerType = itemProvider
-			continue
-		}
-		if providerType != itemProvider {
-			return "", core.NewInvalidRequestError("native batch supports a single provider per batch; split mixed-provider requests", nil)
-		}
-	}
-
-	if providerType == "" {
-		return "", core.NewInvalidRequestError("unable to resolve provider for batch", nil)
-	}
-	return providerType, nil
-}
-
-func mergeBatchRequestEndpointHints(left, right map[string]string) map[string]string {
-	if len(left) == 0 {
-		if len(right) == 0 {
-			return nil
-		}
-		merged := make(map[string]string, len(right))
-		for key, value := range right {
-			merged[key] = value
-		}
-		return merged
-	}
-
-	merged := make(map[string]string, len(left))
-	for key, value := range left {
-		merged[key] = value
-	}
-	for key, value := range right {
-		merged[key] = value
-	}
-	return merged
-}
-
-func (h *Handler) cleanupPreparedBatchInputFile(ctx context.Context, providerType, fileID string) {
-	fileID = strings.TrimSpace(fileID)
-	if fileID == "" {
-		return
-	}
-	files, ok := h.provider.(core.NativeFileRoutableProvider)
-	if !ok {
-		return
-	}
-	if _, err := files.DeleteFile(ctx, providerType, fileID); err != nil {
-		slog.Warn("failed to delete rewritten batch input file", "provider", providerType, "file_id", fileID, "error", err)
-	}
-}
-
-func (h *Handler) loadBatch(c *echo.Context, id string) (*batchstore.StoredBatch, error) {
-	resp, err := h.batchStore.Get(c.Request().Context(), id)
-	if err != nil {
-		if errors.Is(err, batchstore.ErrNotFound) {
-			return nil, core.NewNotFoundError("batch not found: " + id)
-		}
-		return nil, core.NewProviderError("batch_store", http.StatusInternalServerError, "failed to load batch", err)
-	}
-	if resp.Batch != nil {
-		auditlog.EnrichEntry(c, "batch", resp.Batch.Provider)
-	}
-	return resp, nil
+	return h.nativeBatch().Batches(c)
 }
 
 // GetBatch handles GET /v1/batches/{id}.
@@ -1390,55 +360,7 @@ func (h *Handler) loadBatch(c *echo.Context, id string) (*batchstore.StoredBatch
 // @Failure      502  {object}  core.GatewayError
 // @Router       /v1/batches/{id} [get]
 func (h *Handler) GetBatch(c *echo.Context) error {
-	ctx, _ := requestContextWithRequestID(c.Request())
-
-	batchMeta, err := batchRouteInfoFromSemantics(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-	id := ""
-	if batchMeta != nil {
-		id = strings.TrimSpace(batchMeta.BatchID)
-	}
-	if id == "" {
-		return handleError(c, core.NewInvalidRequestError("batch id is required", nil))
-	}
-
-	resp, err := h.loadBatch(c, id)
-	if err != nil {
-		return handleError(c, err)
-	}
-	if resp.Batch == nil {
-		return handleError(c, core.NewProviderError("batch_store", http.StatusInternalServerError, "stored batch payload missing", nil))
-	}
-
-	nativeRouter, ok := h.provider.(core.NativeBatchRoutableProvider)
-	if !ok {
-		return c.JSON(http.StatusOK, resp.Batch)
-	}
-	if resp.Batch.Provider == "" || resp.Batch.ProviderBatchID == "" {
-		return c.JSON(http.StatusOK, resp.Batch)
-	}
-
-	latest, err := nativeRouter.GetBatch(ctx, resp.Batch.Provider, resp.Batch.ProviderBatchID)
-	if err != nil {
-		return handleError(c, err)
-	}
-	updated := false
-	if latest != nil {
-		mergeStoredBatchFromUpstream(resp, latest)
-		updated = true
-	}
-	if isTerminalBatchStatus(resp.Batch.Status) && h.cleanupStoredBatchRewrittenInputFile(ctx, resp) {
-		updated = true
-	}
-	if updated {
-		if err := h.batchStore.Update(c.Request().Context(), resp); err != nil && !errors.Is(err, batchstore.ErrNotFound) {
-			return handleError(c, core.NewProviderError("batch_store", http.StatusInternalServerError, "failed to persist refreshed batch", err))
-		}
-	}
-
-	return c.JSON(http.StatusOK, resp.Batch)
+	return h.nativeBatch().GetBatch(c)
 }
 
 // ListBatches handles GET /v1/batches.
@@ -1456,61 +378,7 @@ func (h *Handler) GetBatch(c *echo.Context) error {
 // @Failure      500    {object}  core.GatewayError
 // @Router       /v1/batches [get]
 func (h *Handler) ListBatches(c *echo.Context) error {
-	batchMeta, err := batchRouteInfoFromSemantics(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-
-	limit := 20
-	if batchMeta != nil && batchMeta.HasLimit {
-		limit = batchMeta.Limit
-	}
-
-	after := ""
-	if batchMeta != nil {
-		after = strings.TrimSpace(batchMeta.After)
-	}
-	normalizedLimit := limit
-	if normalizedLimit <= 0 {
-		normalizedLimit = 20
-	}
-	if normalizedLimit > 100 {
-		normalizedLimit = 100
-	}
-
-	items, err := h.batchStore.List(c.Request().Context(), normalizedLimit+1, after)
-	if err != nil {
-		if errors.Is(err, batchstore.ErrNotFound) {
-			return handleError(c, core.NewNotFoundError("after cursor batch not found: "+after))
-		}
-		return handleError(c, core.NewProviderError("batch_store", http.StatusInternalServerError, "failed to list batches", err))
-	}
-	auditlog.EnrichEntry(c, "batch", "")
-
-	hasMore := len(items) > normalizedLimit
-	if hasMore {
-		items = items[:normalizedLimit]
-	}
-
-	data := make([]core.BatchResponse, 0, len(items))
-	for _, item := range items {
-		if item == nil || item.Batch == nil {
-			continue
-		}
-		data = append(data, *item.Batch)
-	}
-
-	resp := core.BatchListResponse{
-		Object:  "list",
-		Data:    data,
-		HasMore: hasMore,
-	}
-	if len(data) > 0 {
-		resp.FirstID = data[0].ID
-		resp.LastID = data[len(data)-1].ID
-	}
-
-	return c.JSON(http.StatusOK, resp)
+	return h.nativeBatch().ListBatches(c)
 }
 
 // CancelBatch handles POST /v1/batches/{id}/cancel.
@@ -1529,52 +397,7 @@ func (h *Handler) ListBatches(c *echo.Context) error {
 // @Failure      502  {object}  core.GatewayError
 // @Router       /v1/batches/{id}/cancel [post]
 func (h *Handler) CancelBatch(c *echo.Context) error {
-	ctx, _ := requestContextWithRequestID(c.Request())
-
-	batchMeta, err := batchRouteInfoFromSemantics(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-	id := ""
-	if batchMeta != nil {
-		id = strings.TrimSpace(batchMeta.BatchID)
-	}
-	if id == "" {
-		return handleError(c, core.NewInvalidRequestError("batch id is required", nil))
-	}
-
-	resp, err := h.loadBatch(c, id)
-	if err != nil {
-		return handleError(c, err)
-	}
-	if resp.Batch == nil {
-		return handleError(c, core.NewProviderError("batch_store", http.StatusInternalServerError, "stored batch payload missing", nil))
-	}
-
-	nativeRouter, ok := h.provider.(core.NativeBatchRoutableProvider)
-	if !ok || resp.Batch.Provider == "" || resp.Batch.ProviderBatchID == "" {
-		return handleError(c, core.NewInvalidRequestError("native batch cancellation is not available", nil))
-	}
-
-	latest, err := nativeRouter.CancelBatch(ctx, resp.Batch.Provider, resp.Batch.ProviderBatchID)
-	if err != nil {
-		return handleError(c, err)
-	}
-	if latest != nil {
-		mergeStoredBatchFromUpstream(resp, latest)
-	}
-	if isTerminalBatchStatus(resp.Batch.Status) {
-		h.cleanupStoredBatchRewrittenInputFile(ctx, resp)
-	}
-
-	if err := h.batchStore.Update(c.Request().Context(), resp); err != nil {
-		if errors.Is(err, batchstore.ErrNotFound) {
-			return handleError(c, core.NewNotFoundError("batch not found: "+id))
-		}
-		return handleError(c, core.NewProviderError("batch_store", http.StatusInternalServerError, "failed to cancel batch", err))
-	}
-
-	return c.JSON(http.StatusOK, resp.Batch)
+	return h.nativeBatch().CancelBatch(c)
 }
 
 // BatchResults handles GET /v1/batches/{id}/results.
@@ -1593,549 +416,5 @@ func (h *Handler) CancelBatch(c *echo.Context) error {
 // @Failure      502  {object}  core.GatewayError
 // @Router       /v1/batches/{id}/results [get]
 func (h *Handler) BatchResults(c *echo.Context) error {
-	ctx, requestID := requestContextWithRequestID(c.Request())
-
-	batchMeta, err := batchRouteInfoFromSemantics(c)
-	if err != nil {
-		return handleError(c, err)
-	}
-	id := ""
-	if batchMeta != nil {
-		id = strings.TrimSpace(batchMeta.BatchID)
-	}
-	if id == "" {
-		return handleError(c, core.NewInvalidRequestError("batch id is required", nil))
-	}
-
-	stored, err := h.loadBatch(c, id)
-	if err != nil {
-		return handleError(c, err)
-	}
-	if stored.Batch == nil {
-		return handleError(c, core.NewProviderError("batch_store", http.StatusInternalServerError, "stored batch payload missing", nil))
-	}
-
-	nativeRouter, ok := h.provider.(core.NativeBatchRoutableProvider)
-	if !ok || stored.Batch.Provider == "" || stored.Batch.ProviderBatchID == "" {
-		return c.JSON(http.StatusOK, core.BatchResultsResponse{
-			Object:  "list",
-			BatchID: stored.Batch.ID,
-			Data:    stored.Batch.Results,
-		})
-	}
-
-	var upstream *core.BatchResultsResponse
-	if hintedRouter, ok := nativeRouter.(core.NativeBatchHintRoutableProvider); ok && len(stored.RequestEndpointByCustomID) > 0 {
-		upstream, err = hintedRouter.GetBatchResultsWithHints(ctx, stored.Batch.Provider, stored.Batch.ProviderBatchID, stored.RequestEndpointByCustomID)
-	} else {
-		upstream, err = nativeRouter.GetBatchResults(ctx, stored.Batch.Provider, stored.Batch.ProviderBatchID)
-	}
-	if err != nil {
-		if isNativeBatchResultsPending(err) {
-			if latest, getErr := nativeRouter.GetBatch(ctx, stored.Batch.Provider, stored.Batch.ProviderBatchID); getErr == nil && latest != nil {
-				mergeStoredBatchFromUpstream(stored, latest)
-				if updateErr := h.batchStore.Update(c.Request().Context(), stored); updateErr != nil && !errors.Is(updateErr, batchstore.ErrNotFound) {
-					slog.Warn(
-						"failed to update batch store after refreshing pending results",
-						"batch_id", stored.Batch.ID,
-						"provider", stored.Batch.Provider,
-						"provider_batch_id", stored.Batch.ProviderBatchID,
-						"error", updateErr,
-					)
-				}
-			}
-			status := strings.TrimSpace(stored.Batch.Status)
-			if status == "" {
-				status = "in_progress"
-			}
-			return handleError(c, core.NewInvalidRequestErrorWithStatus(
-				http.StatusConflict,
-				fmt.Sprintf("batch results are not ready yet (status: %s)", status),
-				err,
-			))
-		}
-		return handleError(c, err)
-	}
-	if upstream == nil {
-		return handleError(c, core.NewProviderError(stored.Batch.Provider, http.StatusBadGateway, "provider returned empty batch results response", nil))
-	}
-
-	result := *upstream
-	result.BatchID = stored.Batch.ID
-	usageLogged := h.logBatchUsageFromBatchResults(stored, &result, requestID)
-	if len(result.Data) > 0 {
-		stored.Batch.Results = result.Data
-	}
-	cleanedRewrittenInput := false
-	if isTerminalBatchStatus(stored.Batch.Status) {
-		cleanedRewrittenInput = h.cleanupStoredBatchRewrittenInputFile(ctx, stored)
-	}
-	if len(result.Data) > 0 || usageLogged || cleanedRewrittenInput {
-		if updateErr := h.batchStore.Update(c.Request().Context(), stored); updateErr != nil {
-			slog.Warn(
-				"failed to update batch store after receiving batch results",
-				"batch_id", stored.Batch.ID,
-				"provider", stored.Batch.Provider,
-				"provider_batch_id", stored.Batch.ProviderBatchID,
-				"error", updateErr,
-			)
-		}
-	}
-
-	return c.JSON(http.StatusOK, result)
-}
-
-func (h *Handler) logBatchUsageFromBatchResults(stored *batchstore.StoredBatch, result *core.BatchResultsResponse, fallbackRequestID string) bool {
-	if h.usageLogger == nil || !h.usageLogger.Config().Enabled || stored == nil || stored.Batch == nil || result == nil || len(result.Data) == 0 {
-		return false
-	}
-	if stored.UsageLoggedAt != nil {
-		return false
-	}
-
-	requestID := strings.TrimSpace(stored.RequestID)
-	if requestID == "" {
-		requestID = strings.TrimSpace(fallbackRequestID)
-	}
-	if requestID == "" {
-		requestID = "batch:" + stored.Batch.ID
-	}
-
-	loggedEntries := 0
-	inputTotal := 0
-	outputTotal := 0
-	totalTokens := 0
-	var inputCostTotal float64
-	var outputCostTotal float64
-	var totalCostTotal float64
-	hasAnyCost := false
-
-	for _, item := range result.Data {
-		if item.StatusCode < http.StatusOK || item.StatusCode >= http.StatusMultipleChoices {
-			continue
-		}
-
-		payload, ok := asJSONMap(item.Response)
-		if !ok {
-			continue
-		}
-		usagePayload, ok := asJSONMap(payload["usage"])
-		if !ok {
-			continue
-		}
-
-		inputTokens, outputTokens, usageTotal, hasUsage := extractTokenTotals(usagePayload)
-		if !hasUsage {
-			continue
-		}
-
-		provider := firstNonEmpty(item.Provider, stored.Batch.Provider)
-		model := firstNonEmpty(item.Model, stringFromAny(payload["model"]))
-		providerID := firstNonEmpty(
-			stringFromAny(payload["id"]),
-			item.CustomID,
-			fmt.Sprintf("%s:%d", firstNonEmpty(stored.Batch.ProviderBatchID, stored.Batch.ID), item.Index),
-		)
-		rawUsage := buildBatchUsageRawData(usagePayload, stored.Batch, item)
-
-		var pricing *core.ModelPricing
-		if h.pricingResolver != nil && model != "" {
-			pricing = h.pricingResolver.ResolvePricing(model, provider)
-		}
-
-		entry := usage.ExtractFromSSEUsage(
-			providerID,
-			inputTokens,
-			outputTokens,
-			usageTotal,
-			rawUsage,
-			requestID,
-			model,
-			provider,
-			"/v1/batches",
-			pricing,
-		)
-		if entry == nil {
-			continue
-		}
-		entry.ID = deterministicBatchUsageID(stored.Batch, item, providerID)
-
-		h.usageLogger.Write(entry)
-		loggedEntries++
-		inputTotal += inputTokens
-		outputTotal += outputTokens
-		totalTokens += usageTotal
-		if entry.InputCost != nil {
-			inputCostTotal += *entry.InputCost
-			hasAnyCost = true
-		}
-		if entry.OutputCost != nil {
-			outputCostTotal += *entry.OutputCost
-			hasAnyCost = true
-		}
-		if entry.TotalCost != nil {
-			totalCostTotal += *entry.TotalCost
-			hasAnyCost = true
-		}
-	}
-
-	if loggedEntries == 0 {
-		return false
-	}
-
-	now := time.Now().UTC()
-	stored.RequestID = requestID
-	stored.UsageLoggedAt = &now
-
-	stored.Batch.Usage.InputTokens = inputTotal
-	stored.Batch.Usage.OutputTokens = outputTotal
-	stored.Batch.Usage.TotalTokens = totalTokens
-	if hasAnyCost {
-		stored.Batch.Usage.InputCost = &inputCostTotal
-		stored.Batch.Usage.OutputCost = &outputCostTotal
-		stored.Batch.Usage.TotalCost = &totalCostTotal
-	}
-
-	return true
-}
-
-func deterministicBatchUsageID(stored *core.BatchResponse, item core.BatchResultItem, providerID string) string {
-	seed := fmt.Sprintf(
-		"%s|%s|%d|%s|%s",
-		firstNonEmpty(stored.ID, stored.ProviderBatchID),
-		firstNonEmpty(stored.ProviderBatchID, stored.ID),
-		item.Index,
-		item.CustomID,
-		providerID,
-	)
-	return uuid.NewSHA1(uuid.NameSpaceOID, []byte(seed)).String()
-}
-
-func buildBatchUsageRawData(usagePayload map[string]any, stored *core.BatchResponse, item core.BatchResultItem) map[string]any {
-	if usagePayload == nil {
-		return nil
-	}
-
-	raw := make(map[string]any)
-	for key, value := range usagePayload {
-		switch key {
-		case "input_tokens", "output_tokens", "prompt_tokens", "completion_tokens", "total_tokens":
-			continue
-		default:
-			raw[key] = value
-		}
-	}
-
-	if promptDetails, ok := asJSONMap(usagePayload["prompt_tokens_details"]); ok {
-		for key, value := range promptDetails {
-			raw["prompt_"+key] = value
-		}
-	}
-	if completionDetails, ok := asJSONMap(usagePayload["completion_tokens_details"]); ok {
-		for key, value := range completionDetails {
-			raw["completion_"+key] = value
-		}
-	}
-
-	raw["batch_id"] = stored.ID
-	raw["provider_batch_id"] = stored.ProviderBatchID
-	raw["batch_result_index"] = item.Index
-	if item.CustomID != "" {
-		raw["batch_custom_id"] = item.CustomID
-	}
-	if endpoint := strings.TrimSpace(stored.Endpoint); endpoint != "" {
-		raw["batch_endpoint"] = endpoint
-	}
-
-	return raw
-}
-
-func extractTokenTotals(usagePayload map[string]any) (int, int, int, bool) {
-	inputTokens, hasInput := readFirstInt(usagePayload, "input_tokens", "prompt_tokens")
-	outputTokens, hasOutput := readFirstInt(usagePayload, "output_tokens", "completion_tokens")
-	totalTokens, hasTotal := readFirstInt(usagePayload, "total_tokens")
-	if !hasTotal && (hasInput || hasOutput) {
-		totalTokens = inputTokens + outputTokens
-		hasTotal = true
-	}
-
-	return inputTokens, outputTokens, totalTokens, hasInput || hasOutput || hasTotal
-}
-
-func readFirstInt(values map[string]any, keys ...string) (int, bool) {
-	for _, key := range keys {
-		value, exists := values[key]
-		if !exists {
-			continue
-		}
-		if num, ok := intFromAny(value); ok {
-			return num, true
-		}
-	}
-	return 0, false
-}
-
-func intFromAny(value any) (int, bool) {
-	switch v := value.(type) {
-	case int:
-		return v, true
-	case int8:
-		return int(v), true
-	case int16:
-		return int(v), true
-	case int32:
-		return int(v), true
-	case int64:
-		return intFromInt64(v)
-	case uint:
-		return intFromUint64(uint64(v))
-	case uint8:
-		return int(v), true
-	case uint16:
-		return int(v), true
-	case uint32:
-		return intFromUint64(uint64(v))
-	case uint64:
-		return intFromUint64(v)
-	case float32:
-		return intFromFloat64(float64(v))
-	case float64:
-		return intFromFloat64(v)
-	case json.Number:
-		i, err := v.Int64()
-		if err == nil {
-			return intFromInt64(i)
-		}
-		f, err := v.Float64()
-		if err == nil {
-			return intFromFloat64(f)
-		}
-		return 0, false
-	case string:
-		if strings.TrimSpace(v) == "" {
-			return 0, false
-		}
-		if i, err := strconv.Atoi(strings.TrimSpace(v)); err == nil {
-			return i, true
-		}
-		return 0, false
-	default:
-		return 0, false
-	}
-}
-
-func intFromInt64(v int64) (int, bool) {
-	maxInt := int64(^uint(0) >> 1)
-	minInt := -maxInt - 1
-	if v < minInt || v > maxInt {
-		return 0, false
-	}
-	return int(v), true
-}
-
-func intFromUint64(v uint64) (int, bool) {
-	maxInt := uint64(^uint(0) >> 1)
-	if v > maxInt {
-		return 0, false
-	}
-	return int(v), true
-}
-
-func intFromFloat64(v float64) (int, bool) {
-	if math.IsNaN(v) || math.IsInf(v, 0) {
-		return 0, false
-	}
-	maxInt := float64(^uint(0) >> 1)
-	minInt := -maxInt - 1
-	if v < minInt || v > maxInt {
-		return 0, false
-	}
-	return int(v), true
-}
-
-func asJSONMap(value any) (map[string]any, bool) {
-	switch v := value.(type) {
-	case map[string]any:
-		return v, true
-	case json.RawMessage:
-		return decodeJSONMap(v)
-	case []byte:
-		return decodeJSONMap(v)
-	case string:
-		if strings.TrimSpace(v) == "" {
-			return nil, false
-		}
-		return decodeJSONMap([]byte(v))
-	default:
-		raw, err := json.Marshal(v)
-		if err != nil {
-			return nil, false
-		}
-		return decodeJSONMap(raw)
-	}
-}
-
-func decodeJSONMap(raw []byte) (map[string]any, bool) {
-	if len(raw) == 0 {
-		return nil, false
-	}
-	var decoded map[string]any
-	if err := json.Unmarshal(raw, &decoded); err != nil {
-		return nil, false
-	}
-	return decoded, true
-}
-
-func stringFromAny(value any) string {
-	switch v := value.(type) {
-	case string:
-		return strings.TrimSpace(v)
-	case []byte:
-		return strings.TrimSpace(string(v))
-	default:
-		return ""
-	}
-}
-
-func firstNonEmpty(values ...string) string {
-	for _, value := range values {
-		if trimmed := strings.TrimSpace(value); trimmed != "" {
-			return trimmed
-		}
-	}
-	return ""
-}
-
-func mergeStoredBatchFromUpstream(stored *batchstore.StoredBatch, upstream *core.BatchResponse) {
-	if stored == nil || stored.Batch == nil || upstream == nil {
-		return
-	}
-
-	stored.Batch.Status = firstNonEmpty(upstream.Status, stored.Batch.Status)
-	stored.Batch.Endpoint = firstNonEmpty(upstream.Endpoint, stored.Batch.Endpoint)
-	if strings.TrimSpace(stored.Batch.InputFileID) == "" {
-		stored.Batch.InputFileID = firstNonEmpty(stored.OriginalInputFileID, upstream.InputFileID)
-	}
-	stored.Batch.CompletionWindow = firstNonEmpty(upstream.CompletionWindow, stored.Batch.CompletionWindow)
-	if hasBatchRequestCounts(upstream.RequestCounts) {
-		stored.Batch.RequestCounts = upstream.RequestCounts
-	}
-	if hasBatchUsageSummary(upstream.Usage) {
-		stored.Batch.Usage = upstream.Usage
-	}
-	if len(upstream.Results) > 0 {
-		stored.Batch.Results = upstream.Results
-	}
-	if upstream.InProgressAt != nil {
-		stored.Batch.InProgressAt = upstream.InProgressAt
-	}
-	if upstream.CompletedAt != nil {
-		stored.Batch.CompletedAt = upstream.CompletedAt
-	}
-	if upstream.FailedAt != nil {
-		stored.Batch.FailedAt = upstream.FailedAt
-	}
-	if upstream.CancellingAt != nil {
-		stored.Batch.CancellingAt = upstream.CancellingAt
-	}
-	if upstream.CancelledAt != nil {
-		stored.Batch.CancelledAt = upstream.CancelledAt
-	}
-	if upstream.Metadata != nil {
-		if stored.Batch.Metadata == nil {
-			stored.Batch.Metadata = map[string]string{}
-		}
-		preservedGatewayMetadata := map[string]string{}
-		for _, key := range []string{"provider", "provider_batch_id"} {
-			if value, exists := stored.Batch.Metadata[key]; exists {
-				preservedGatewayMetadata[key] = value
-			}
-		}
-		for key, value := range upstream.Metadata {
-			if _, preserve := preservedGatewayMetadata[key]; preserve {
-				continue
-			}
-			stored.Batch.Metadata[key] = value
-		}
-		for key, value := range preservedGatewayMetadata {
-			stored.Batch.Metadata[key] = value
-		}
-		stored.Batch.Metadata = sanitizePublicBatchMetadata(stored.Batch.Metadata)
-	}
-}
-
-func hasBatchRequestCounts(counts core.BatchRequestCounts) bool {
-	return counts.Total != 0 || counts.Completed != 0 || counts.Failed != 0
-}
-
-func hasBatchUsageSummary(usage core.BatchUsageSummary) bool {
-	return usage.InputTokens != 0 ||
-		usage.OutputTokens != 0 ||
-		usage.TotalTokens != 0 ||
-		usage.InputCost != nil ||
-		usage.OutputCost != nil ||
-		usage.TotalCost != nil
-}
-
-func isTerminalBatchStatus(status string) bool {
-	switch strings.ToLower(strings.TrimSpace(status)) {
-	case "completed", "failed", "cancelled", "canceled", "expired":
-		return true
-	default:
-		return false
-	}
-}
-
-func (h *Handler) cleanupStoredBatchRewrittenInputFile(ctx context.Context, stored *batchstore.StoredBatch) bool {
-	if stored == nil || stored.Batch == nil {
-		return false
-	}
-	fileID := strings.TrimSpace(stored.RewrittenInputFileID)
-	if fileID == "" {
-		return false
-	}
-	nativeFiles, err := h.nativeFileRouter()
-	if err != nil {
-		return false
-	}
-	if _, err := nativeFiles.DeleteFile(ctx, stored.Batch.Provider, fileID); err != nil {
-		slog.Warn(
-			"failed to delete rewritten batch input file",
-			"batch_id", stored.Batch.ID,
-			"provider", stored.Batch.Provider,
-			"file_id", fileID,
-			"error", err,
-		)
-		return false
-	}
-	stored.RewrittenInputFileID = ""
-	return true
-}
-
-func isNativeBatchResultsPending(err error) bool {
-	var gatewayErr *core.GatewayError
-	if !errors.As(err, &gatewayErr) {
-		return false
-	}
-	if gatewayErr.HTTPStatusCode() != http.StatusNotFound {
-		return false
-	}
-	// Some providers return 404 while native results are still being prepared.
-	// Extend batchResultsPending404Providers as more provider-specific behaviors are confirmed.
-	_, ok := batchResultsPending404Providers[strings.ToLower(strings.TrimSpace(gatewayErr.Provider))]
-	return ok
-}
-
-// handleError converts gateway errors to appropriate HTTP responses
-func handleError(c *echo.Context, err error) error {
-	var gatewayErr *core.GatewayError
-	if errors.As(err, &gatewayErr) {
-		auditlog.EnrichEntryWithError(c, string(gatewayErr.Type), gatewayErr.Message)
-		return c.JSON(gatewayErr.HTTPStatusCode(), gatewayErr.ToJSON())
-	}
-
-	// Fallback for unexpected errors
-	gatewayErr = core.NewProviderError("", http.StatusInternalServerError, "an unexpected error occurred", err)
-	auditlog.EnrichEntryWithError(c, string(gatewayErr.Type), gatewayErr.Message)
-	return c.JSON(gatewayErr.HTTPStatusCode(), gatewayErr.ToJSON())
+	return h.nativeBatch().BatchResults(c)
 }
