+++ internal/server/http.go
@@ -32,28 +32,29 @@ type Server struct {
 
 // Config holds server configuration options
 type Config struct {
-	MasterKey                   string                                 // Optional: Master key for authentication
-	MetricsEnabled              bool                                   // Whether to expose Prometheus metrics endpoint
-	MetricsEndpoint             string                                 // HTTP path for metrics endpoint (default: /metrics)
-	BodySizeLimit               string                                 // Max request body size (e.g., "10M", "1024K")
-	AuditLogger                 auditlog.LoggerInterface               // Optional: Audit logger for request/response logging
-	UsageLogger                 usage.LoggerInterface                  // Optional: Usage logger for token tracking
-	PricingResolver             usage.PricingResolver                  // Optional: Resolves pricing for cost calculation
-	ModelResolver               RequestModelResolver                   // Optional: explicit model resolver used during request planning
-	TranslatedRequestPatcher    TranslatedRequestPatcher               // Optional: request patcher for translated routes after planning
-	BatchRequestPreparer        BatchRequestPreparer                   // Optional: batch request preparer before native provider submission
-	ExposedModelLister          ExposedModelLister                     // Optional: additional public models to merge into GET /v1/models
-	BatchStore                  batchstore.Store                       // Optional: Batch lifecycle persistence store
-	LogOnlyModelInteractions    bool                                   // Only log AI model endpoints (default: true)
-	DisablePassthroughRoutes    bool                                   // Disable /p/{provider}/{endpoint} route registration
-	EnabledPassthroughProviders []string                               // Provider types enabled on /p/{provider}/... passthrough routes
-	AllowPassthroughV1Alias     *bool                                  // Allow /p/{provider}/v1/... aliases; nil defaults to true
-	AdminEndpointsEnabled       bool                                   // Whether admin API endpoints are enabled
-	AdminUIEnabled              bool                                   // Whether admin dashboard UI is enabled
-	AdminHandler                *admin.Handler                         // Admin API handler (nil if disabled)
-	DashboardHandler            *dashboard.Handler                     // Dashboard UI handler (nil if disabled)
-	SwaggerEnabled              bool                                   // Whether to expose the Swagger UI at /swagger/index.html
-	ResponseCacheMiddleware     *responsecache.ResponseCacheMiddleware // Optional: response cache middleware for cacheable endpoints
+	MasterKey                    string                                 // Optional: Master key for authentication
+	MetricsEnabled               bool                                   // Whether to expose Prometheus metrics endpoint
+	MetricsEndpoint              string                                 // HTTP path for metrics endpoint (default: /metrics)
+	BodySizeLimit                string                                 // Max request body size (e.g., "10M", "1024K")
+	AuditLogger                  auditlog.LoggerInterface               // Optional: Audit logger for request/response logging
+	UsageLogger                  usage.LoggerInterface                  // Optional: Usage logger for token tracking
+	PricingResolver              usage.PricingResolver                  // Optional: Resolves pricing for cost calculation
+	ModelResolver                RequestModelResolver                   // Optional: explicit model resolver used during request planning
+	TranslatedRequestPatcher     TranslatedRequestPatcher               // Optional: request patcher for translated routes after planning
+	BatchRequestPreparer         BatchRequestPreparer                   // Optional: batch request preparer before native provider submission
+	ExposedModelLister           ExposedModelLister                     // Optional: additional public models to merge into GET /v1/models
+	PassthroughSemanticEnrichers []core.PassthroughSemanticEnricher     // Optional: provider-owned passthrough semantic enrichers before planning
+	BatchStore                   batchstore.Store                       // Optional: Batch lifecycle persistence store
+	LogOnlyModelInteractions     bool                                   // Only log AI model endpoints (default: true)
+	DisablePassthroughRoutes     bool                                   // Disable /p/{provider}/{endpoint} route registration
+	EnabledPassthroughProviders  []string                               // Provider types enabled on /p/{provider}/... passthrough routes
+	AllowPassthroughV1Alias      *bool                                  // Allow /p/{provider}/v1/... aliases; nil defaults to true
+	AdminEndpointsEnabled        bool                                   // Whether admin API endpoints are enabled
+	AdminUIEnabled               bool                                   // Whether admin dashboard UI is enabled
+	AdminHandler                 *admin.Handler                         // Admin API handler (nil if disabled)
+	DashboardHandler             *dashboard.Handler                     // Dashboard UI handler (nil if disabled)
+	SwaggerEnabled               bool                                   // Whether to expose the Swagger UI at /swagger/index.html
+	ResponseCacheMiddleware      *responsecache.ResponseCacheMiddleware // Optional: response cache middleware for cacheable endpoints
 }
 
 // New creates a new HTTP server
@@ -191,6 +192,10 @@ func New(provider core.RoutableProvider, cfg *Config) *Server {
 	// Authentication (skips public paths)
 	if cfg != nil && cfg.MasterKey != "" {
 		e.Use(AuthMiddleware(cfg.MasterKey, authSkipPaths))
+	}
+
+	if cfg != nil && len(cfg.PassthroughSemanticEnrichers) > 0 {
+		e.Use(PassthroughSemanticEnrichment(cfg.PassthroughSemanticEnrichers, passthroughV1PrefixNormalizationEnabled(cfg)))
 	}
 
 	// Request planning (skips non-model paths via IsModelInteractionPath)
