+++ internal/server/request_snapshot_test.go
@@ -107,6 +107,11 @@ func TestRequestSnapshotCapture_PreservesPassthroughRouteParams(t *testing.T) {
 	assert.Equal(t, "provider_passthrough", capturedEnv.RouteType)
 	assert.Equal(t, "openai", capturedEnv.RouteHints.Provider)
 	assert.Equal(t, "responses", capturedEnv.RouteHints.Endpoint)
+	if info := capturedEnv.CachedPassthroughRouteInfo(); assert.NotNil(t, info) {
+		assert.Equal(t, "openai", info.Provider)
+		assert.Equal(t, "responses", info.RawEndpoint)
+		assert.Equal(t, "/p/openai/responses", info.AuditPath)
+	}
 }
 
 func TestRequestSnapshotCapture_GeneratesRequestIDWhenMissing(t *testing.T) {
