+++ internal/server/stream_support.go
@@ -1,0 +1,32 @@
+package server
+
+import (
+	"io"
+	"net/http"
+)
+
+func flushStream(w io.Writer, stream io.Reader) error {
+	flusher, canFlush := w.(http.Flusher)
+	if canFlush {
+		flusher.Flush()
+	}
+
+	buf := make([]byte, 32*1024)
+	for {
+		n, err := stream.Read(buf)
+		if n > 0 {
+			if _, writeErr := w.Write(buf[:n]); writeErr != nil {
+				return writeErr
+			}
+			if canFlush {
+				flusher.Flush()
+			}
+		}
+		if err != nil {
+			if err == io.EOF {
+				return nil
+			}
+			return err
+		}
+	}
+}
